+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] - 2021-03-04
+
+* Listener: Override attributes when reading them from the environment variables
+* Listener: Delete a queue
+* Listener: Stop a consumer
+
 ## [1.0.2] - 2021-02-26
 
 * Allow users to define a RabbitMQExceptionHandler without using reflection